trying out moving the function definition to the source file
