+++ manifest.ts
@@ -1,34 +1,5 @@
-import {
-  DefineFunction,
-  DefineWorkflow,
-  Manifest,
-  Schema,
-} from "deno-slack-sdk/mod.ts";
-
-export const ReverseFunction = DefineFunction({
-  callback_id: "reverse",
-  title: "Reverse",
-  description: "Takes a string and reverses it",
-  source_file: "functions/reverse.ts",
-  input_parameters: {
-    properties: {
-      stringToReverse: {
-        type: Schema.types.string,
-        description: "The string to reverse",
-      },
-    },
-    required: ["stringToReverse"],
-  },
-  output_parameters: {
-    properties: {
-      reverseString: {
-        type: Schema.types.string,
-        description: "The string in reverse",
-      },
-    },
-    required: ["reverseString"],
-  },
-});
+import { DefineWorkflow, Manifest, Schema } from "deno-slack-sdk/mod.ts";
+import { ReverseFunction } from "./functions/reverse.ts";
 
 export const TestReverseWorkflow = DefineWorkflow({
   callback_id: "test_reverse",
